updated comments and examples

+++ shared_mem_dict.py
@@ -135,7 +135,7 @@ class SharedMemoryConfigurationError(Exception):
     pass
 
 class SharedMemDict():
-    def __init__(self,name,num,dtype,reset_shm=False,varnames=[],verbose=False):
+    def __init__(self,name,num,dtype,reset_shm=False,varnames=[],check_config=True,verbose=False):
         """Create or connect to a shared memory dict. Values in the dict can be accessed by name or
         by numerical index; the values are all a single data type.
 
@@ -156,6 +156,15 @@ class SharedMemDict():
             A list of names for each shared memory array element. This is used to provide dict-style
             access to individual elements. The default is None. The length of this list must be
             equal to num.
+        check_config : bool, Optional
+            If True, calculate the sha256 hash of the configuration and compare
+            it to the hash stored in a separate block of the shared memory
+            area. If the hashes differ, raise the SharedMemoryConfigurationError.
+            The default is True (always check). This should only be disabled if
+            debugging is required.
+        verbose : bool, Optional
+            If True, print additional information during object creation. The 
+            default is False.
 
         Returns
         -------
@@ -219,29 +228,33 @@ class SharedMemDict():
         self.nbytes_cfghash = len(my_cfg_sha256)
         self.nbytes = self.nbytes_data + self.nbytes_cfghash
 
-        # if no names for the variables are provided, the index numbers are used
+        # If no variable names were supplied, make the default variable names
+        # consisting of a list of int starting at 0.
         try:
             if varnames is None or len(varnames)==0:
                 varnames = list(range(num))
         except TypeError as e:
             estr = f"'varnames' must be an iterable type (list, tuple) of length 0 or {num} ('num' argument)."
             raise Exception(estr).with_traceback(e.__traceback__)
-
+            
+        # Check for mismatch between length of varirable names and expected number of names
         if len(varnames) != self.num:
-            raise ValueError(f"Number of variable names ({len(varnames)}) do not match num given in init call ({num}).")
+            raise ValueError(f"Number of variable names (len(varnames)={len(varnames)}) do not match expected number (num={num}).")
 
+        # Check for duplicate names
         if len(varnames) != len(set(varnames)):
             [varnames.pop(varnames.index(k)) for k in set(varnames)] # pop all names once; leftovers are the duplicates
             raise ValueError(f"Found repeated variable names: {set(varnames)}")
 
+        # Create the variable reference dict (name:index)
         self.varnames = {var:i for var,i in zip(varnames,range(num))}
 
-        # create the shared memory
+        # Create the shared memory
         try:
             self.shm = shared_memory.SharedMemory(name=self.name,create=True,size=self.nbytes)
             self.shm.buf[:] = bytearray(self.nbytes) # init to zeros if we are creating
         except FileExistsError:
-            if reset_shm:
+            if reset_shm and verbose:
                 print(f'    shm reset: {self.name}')
                 tmp = shared_memory.SharedMemory(name=self.name)
                 tmp.unlink() # Calling unlink here causes all those "cannot close" errors
@@ -251,23 +264,26 @@ class SharedMemDict():
         self.arr = self.shm.buf[:self.nbytes_data].cast(self.fmt)
         self._cfg = self.shm.buf[self.nbytes_data:(self.nbytes_data+self.nbytes_cfghash)].cast('B')
 
-
-        all_zero = all([b==0 for b in self._cfg])
-        cfg_match = all([self._cfg[i]==my_cfg_sha256[i] for i in range(self.nbytes_cfghash)])
-        if all_zero:
-            if verbose:
-                print(f"Config check area was all zeros. Writing my_cfg_sha256: {my_cfg_sha256.hex()}")
-            self._cfg[:self.nbytes_cfghash] = my_cfg_sha256
-        elif not cfg_match:
-            estr = f"Shared Memory configuration does not match source. Check that the " \
-                   f"configuration is identical (order and case). " \
-                   f" "
-                       
-                       "The configuration used by" \
-                   f" this instance is:\n\n{my_cfg}"
-            self.close()
-            raise SharedMemoryConfigurationError(estr)
-
+        # Check for a matching configuration hash
+        if check_config:
+            all_zero = all([b==0 for b in self._cfg])
+            cfg_match = all([self._cfg[i]==my_cfg_sha256[i] for i in range(self.nbytes_cfghash)])
+            if all_zero:
+                if verbose:
+                    print(f"Config check area was all zeros. Writing my_cfg_sha256: {my_cfg_sha256.hex()}")
+                self._cfg[:self.nbytes_cfghash] = my_cfg_sha256
+            elif cfg_match:
+                if verbose:
+                    print("Config hash matched to existing hash.")
+            else:
+                estr = f"    Shared Memory configuration does not match source configuration. " \
+                        "Check that the configuration used by this instance is identical " \
+                        "in variable name, case, and order. The configuration " \
+                       f"used by this instance is:\n\n{my_cfg}\n\n" \
+                       f"    This instance's config sha256 hash is: {my_cfg_sha256.hex()}\n" \
+                       f"    The existing config sha256 hash is:    {self._cfg.hex()}" 
+                self.close()
+                raise SharedMemoryConfigurationError(estr)
 
         # # connect shared memory buffer to a numpy ndarray obj
         # if no_numpy:
@@ -298,17 +314,19 @@ class SharedMemDict():
         return self.varnames.keys()
 
     def values(self):
-        return [self.arr[i] for i in range(self.num)]  # TODO: why can arr be longer than num?
+        return list(self.arr[:self.num])
+        # return [self.arr[i] for i in range(self.num)]  # NOTE: arr be longer than num
 
     def items(self):
         for key in self.keys():
             yield ( key, self.arr[self.varnames[key]])
 
-    def getvar(self,varname):
-        return self.arr[ self.varnames[varname] ]
+    def getvar(self,key):
+        return self.arr[ self.varnames[key] ]
 
-    def setvar(self,varname,value):
-        self[varname] = value
+    def setvar(self,key,value):
+        # self[key] = value  # this is slow
+        self.arr[ self.varnames[key] ] = value
 
     def close(self):
         # To prevent a memoryview error message (cannot close exported pointers exist), arr is
@@ -321,12 +339,17 @@ class SharedMemDict():
         self.close()
         self.shm.unlink()
 
+#%% 
 def _demo_process(cfg):
     import time
-    """This is defined here to avoid an error when running the example in iPython."""
+    """Demo: Increment the 7th item in a shared list. This is defined here to 
+    avoid an error when running the example in iPython."""
+    
+    # Connect to the pre-defined shared instance
     shm = SharedMemDict(**cfg)
     n=400
-    print(shm.my_cfg_sha256)
+    # print(shm.my_cfg_sha256)
+    # Increment 
     for i in range(n):
         shm[7] = float(i)
         # print(f"{shm[7]}")
@@ -341,17 +364,14 @@ if __name__ == "__main__":
     from multiprocessing import Process
     import time
 
-
-
     # Example 1: unnammed array
-    print("\n\nExample 1: unnammed array")
+    print("\n\n===========================================")
+    print("Example 1: unnammed array")
     shm_cfg = {'name': 'shm_area_test1_8u235',  # shared memory location identifier - can be anything you want
                'num':  10,  # number of variables in the shared memory array
                'dtype': 'float64',  # datatype of the shared memory array
-               'varnames': [], # optional names for each variable
-               # 'varnames': None, # optional names for each variable
+               'varnames': [], # optional names for each variable - if empty, indices count up from 0
                }
-
 
     shm = SharedMemDict(**shm_cfg)
     print("    Set array items by index number.")
@@ -363,15 +383,20 @@ if __name__ == "__main__":
     print(f"    Dict key:    shm[0]        -> {shm[0]}")
     print(f"    Array Index: shm.arr[0]    -> {shm.arr[0]}")
     print(f"    getvar():    shm.getvar(0) -> {shm.getvar(0)} (equivalent to dict key)")
-    print(f" Get all values:  shm.values(): {shm.values()}")
     print(f" List all keys:   shm.keys(): {shm.keys()}")
+    print(f" Get all values:  shm.values(): {shm.values()}")
+    print(f"    getvar() : {[shm.getvar(i) for i in range(shm.num)]}")
+    print(f" __getitem__ : {[shm[i] for i in range(shm.num)]}")
+    print(f" array index : {[shm.arr[i] for i in range(shm.num)]}")
     # print(shm.my_cfg_sha256)
 
+    print("\n    Spawning demo_process() in a new process to write values to shared memory...")
     p = Process(target=_demo_process,args=(shm_cfg,))
     p.start()
     time.sleep(1.0)
+    print("    Reading values from shared memory:")
     for i in range(10):
-        print(f"Shared from demo_process: shm[7] = {shm[7]}")
+        print(f"        Shared from demo_process: shm[7] = {shm[7]}")
         time.sleep(0.25)
     p.join()
     # %timeit shm[1]
@@ -390,7 +415,8 @@ if __name__ == "__main__":
 
 
     # Example 2: Named variables
-    print("\n\nExample 2: Named Variables")
+    print("\n\n===========================================")
+    print("Example 2: Named Variables")
     shm_cfg = {'name': 'shm_area_test1_8u235',  # shared memory location identifier - can be anything you want
                'num':  10,  # number of variables in the shared memory array
                'dtype': 'float64',  # datatype of the shared memory array
@@ -412,7 +438,8 @@ if __name__ == "__main__":
 
 
     # Example 3 - config mismatch
-    print("\n\nExample 3: Configuration Error Test")
+    print("\n\n===========================================")
+    print("Example 3: Configuration Error Test")
     shm_cfg = {'name': 'shm_area_test1_8u235',  # shared memory location identifier - can be anything you want
                'num':  10,  # number of variables in the shared memory array
                'dtype': 'float64',  # datatype of the shared memory array
